concept-api: Move migrations inside same package as api
