+++ concept-api/src/main/resources/no/ndla/conceptapi/db/migration/V1__Create_Initial_Tables.sql
