+++ concept-api/src/main/resources/no/ndla/conceptapi/db/migration/V2__Add_Listing_Id_Column.sql
