+++ concept-api/src/main/resources/no/ndla/conceptapi/db/migration/V4__Create_Published_Tables.sql
