+++ concept-api/src/test/scala/no/ndla/conceptapi/db/migration/R__MetaImageAsVisualElementTest.scala
@@ -5,9 +5,9 @@
  * See LICENSE
  */
 
-package conceptapi.db.migration
+package no.ndla.conceptapi.db.migration
 
-import conceptapi.db.migration.R__MetaImageAsVisualElement
+import no.ndla.conceptapi.db.migration.R__MetaImageAsVisualElement
 import no.ndla.conceptapi.{TestEnvironment, UnitSuite}
 
 class R__MetaImageAsVisualElementTest extends UnitSuite with TestEnvironment {
