+++ concept-api/src/test/scala/no/ndla/conceptapi/db/migration/V5__MigrateStatusToQualityAssuredTest.scala
@@ -5,9 +5,9 @@
  * See LICENSE
  */
 
-package conceptapi.db.migration
+package no.ndla.conceptapi.db.migration
 
-import conceptapi.db.migration.V5__MigrateStatusToQualityAssured
+import no.ndla.conceptapi.db.migration.V5__MigrateStatusToQualityAssured
 import no.ndla.conceptapi.{TestEnvironment, UnitSuite}
 
 class V5__MigrateStatusToQualityAssuredTest extends UnitSuite with TestEnvironment {
