+++ concept-api/src/test/scala/no/ndla/conceptapi/db/migration/V8__PublishedConceptArticleIdsAsListTest.scala
@@ -5,9 +5,9 @@
  * See LICENSE
  */
 
-package conceptapi.db.migration
+package no.ndla.conceptapi.db.migration
 
-import conceptapi.db.migration.V8__PublishedConceptArticleIdsAsList
+import no.ndla.conceptapi.db.migration.V8__PublishedConceptArticleIdsAsList
 import no.ndla.conceptapi.{TestEnvironment, UnitSuite}
 
 class V8__PublishedConceptArticleIdsAsListTest extends UnitSuite with TestEnvironment {
