+++ concept-api/src/test/scala/no/ndla/conceptapi/db/migration/V9__LanguageUnknownToUndTest.scala
@@ -5,9 +5,9 @@
  * See LICENSE
  */
 
-package conceptapi.db.migration
+package no.ndla.conceptapi.db.migration
 
-import conceptapi.db.migration.V9__LanguageUnknownToUnd
+import no.ndla.conceptapi.db.migration.V9__LanguageUnknownToUnd
 import no.ndla.conceptapi.{TestEnvironment, UnitSuite}
 
 class V9__LanguageUnknownToUndTest extends UnitSuite with TestEnvironment {
